+++ src/games/nibbler/src/Terrain.cpp
@@ -16,9 +16,9 @@ Terrain::Terrain(size_t height, size_t width, IGraphicsLib **gfx)
 {
     this->_map = std::vector<std::vector<char>>(height);
 
-    for (int i = 0; i < height; i++) {
+    for (size_t i = 0; i < height; i++) {
         this->_map[i] = std::vector<char>(width);
-        for (int j = 0; j < width; j++) {
+        for (size_t j = 0; j < width; j++) {
             if (i == 0 || i == (height - 1) || j == 0 || (j == width - 1))
                 this->_map[i][j] = TERRAIN_WALL;
             else
@@ -30,7 +30,7 @@ Terrain::Terrain(size_t height, size_t width, IGraphicsLib **gfx)
     this->_walkable.push_back(TERRAIN_FLOOR);
 }
 
-bool Terrain::validLocation(int x, int y)
+bool Terrain::validLocation(size_t x, size_t y)
 {
     if (!((y < (this->_map.size() - 1)) || (x < (this->_map[0].size() - 1))))
         return false;
@@ -42,8 +42,8 @@ bool Terrain::validLocation(int x, int y)
 
 void Terrain::draw()
 {
-    for (int x = 0; x < this->_map.size(); x++) {
-        for (int y = 0; y < this->_map[x].size(); y++) {
+    for (size_t x = 0; x < this->_map.size(); x++) {
+        for (size_t y = 0; y < this->_map[x].size(); y++) {
             GFX->drawTile(this->_map[x][y], x, y);
         }
     }
